Move selection classes into separate files in selections subfolder
