+++ client/component/SelectionConfig.tsx
@@ -2,7 +2,8 @@ import { ComponentType, Fragment, SVGAttributes } from 'react';
 import { HexColorPicker as Picker } from 'react-colorful';
 import { Modeless } from './Modeless';
 import { LabelledInput } from './LabelledInput';
-import { getSelectionLabel, BaseSelection } from './selections';
+import BaseSelection from './selections/BaseSelection';
+import { getSelectionLabel } from './selections/utils';
 import { isNumber, isValidPositiveNumber } from './utils';
 import styles from './SelectionConfig.module.css';
 import { Btn } from '@h5web/lib';
